+++ SourceCgit/SourceCgit.php
@@ -93,7 +93,7 @@ class SourceCgitPlugin extends MantisSourceGitBasePlugin {
 		if ( isset( $p_repo->info['master_branch'] ) ) {
 			$t_master_branch = $p_repo->info['master_branch'];
 		} else {
-			$t_master_branch = 'master';
+			$t_master_branch = $this->get_default_primary_branches();
 		}
 ?>
 <tr>
@@ -157,7 +157,7 @@ class SourceCgitPlugin extends MantisSourceGitBasePlugin {
 
 		$t_branch = $p_repo->info['master_branch'];
 		if ( is_blank( $t_branch ) ) {
-			$t_branch = 'master';
+			$t_branch = $this->get_default_primary_branches();
 		}
 
 		$t_branches = array_map( 'trim', explode( ',', $t_branch ) );